workspace/skills/curate/scripts/requirements.txt
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
psycopg2-binary
feedparser
